Conflicts:
	.gitignore
	ChangeLog

+++ mp.c
@@ -519,7 +519,7 @@ mp-oprs (%s:%d): quit.\n"),
 void send_strings_message_string_socket(int socket, PString rec, PString message )
 {
      int h_size_name,i,total_size, h_size_mess;
-     u_long n_size_name, n_size_mess;                  /* for network */
+     uint32_t n_size_name, n_size_mess;                  /* for network */
 
      char *buf, *tmp;
 
@@ -642,7 +642,7 @@ void add_buffered_message(Mp_Oprs_Client *rec, Mp_Oprs_Client *sender, PString s
 	  
      } else {
 	  int h_size_name, h_size_mess;
-	  u_long n_size_name, n_size_mess;                  /* for network */
+	  uint32_t n_size_name, n_size_mess;                  /* for network */
 	  char *tmp;
 
 	  h_size_name = strlen(sender_name);
